requirements.txt
# ===========================================================================
# Pharma Pricing Copilot — Python Dependencies
# ===========================================================================
# Install with:  pip install -r requirements.txt
# Recommended Python version: 3.11+
# ===========================================================================

# ---------------------------------------------------------------------------
# Anthropic SDK (Claude models + Voyage embeddings)
# ---------------------------------------------------------------------------
anthropic>=0.40.0

# ---------------------------------------------------------------------------
# Data science core
# ---------------------------------------------------------------------------
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0

# ---------------------------------------------------------------------------
# Machine learning (anomaly detection)
# ---------------------------------------------------------------------------
scikit-learn>=1.5.0

# ---------------------------------------------------------------------------
# Vector store (optional but recommended for RAG)
# ---------------------------------------------------------------------------
faiss-cpu>=1.8.0

# ---------------------------------------------------------------------------
# Streamlit app
# ---------------------------------------------------------------------------
streamlit>=1.36.0
plotly>=5.22.0

# ---------------------------------------------------------------------------
# Notebook / EDA
# ---------------------------------------------------------------------------
jupyter>=1.0.0
ipykernel>=6.29.0
matplotlib>=3.9.0
seaborn>=0.13.0

# ---------------------------------------------------------------------------
# PDF ingestion (for regulatory documents)
# ---------------------------------------------------------------------------
pypdf>=4.3.0

# ---------------------------------------------------------------------------
# Environment & configuration
# ---------------------------------------------------------------------------
python-dotenv>=1.0.0

# ---------------------------------------------------------------------------
# Development & testing
# ---------------------------------------------------------------------------
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
black>=24.4.0
ruff>=0.5.0
mypy>=1.10.0
